requirements.txt
boto3
pandas
sqlalchemy
psycopg2-binary
python-dotenv
great_expectations
plotly
dash
requests
pydantic